+++ nebuchadnezzar/nebu/tests/cli/test_get.py
@@ -83,6 +83,95 @@ class TestGetCmd:
                                 pathlib_walk(expected))
         assert sorted(relative_dir) == sorted(relative_expected)
 
+    def test_three_part_vers(self, datadir, tmpcwd, monkeypatch, requests_mock,
+                             invoker):
+        col_id = 'col11405'
+        col_version = '1.1'
+        trip_ver = '1.1.1'
+        col_uuid = 'b699648f-405b-429f-bf11-37bad4246e7c'
+        col_hash = '{}@{}'.format(col_uuid, '1.4')
+        base_url = 'https://archive.cnx.org'
+        metadata_url = '{}/content/{}/{}'.format(base_url, col_id, col_version)
+        extras_url = '{}/extras/{}'.format(base_url, col_hash)
+        extra_extras_url = '{}/extras/{}@{}'.format(base_url, col_uuid, '1.1')
+
+        # Register the data urls
+        for fname, url in (('contents_1.4.json', metadata_url),
+                           ('extras_1.4.json', extras_url),
+                           ('extras_old.json', extra_extras_url),
+                           ):
+            register_data_file(requests_mock, datadir, fname, url)
+
+        # Register the resources
+        resdir = datadir / 'resources'
+        for res in resdir.glob('*'):
+            url = '{}/resources/{}'.format(base_url, res.relative_to(resdir))
+            register_data_file(requests_mock, resdir, res, url)
+
+        # Register contents
+        condir = datadir / 'contents'
+        for con in condir.glob('*'):
+            url = '{}/contents/{}'.format(base_url, con.relative_to(condir))
+            register_data_file(requests_mock, condir, con, url)
+
+        # Register subcollection/chapter as 404
+        register_404(requests_mock,
+                     'https://archive.cnx.org/contents/'
+                     '8ddfc8de-5164-5828-9fed-d0ed17edb489@1.1')
+
+        # patch input to return 'y' - getting not-head
+        with monkeypatch.context() as m:
+            m.setattr('builtins.input', lambda x: "y")
+            from nebu.cli.main import cli
+            args = ['get', 'test-env', col_id, trip_ver]
+            result = invoker(cli, args)
+
+        assert result.exit_code == 0
+
+        dir = tmpcwd / '{}_1.{}'.format(col_id, '1.1')
+        expected = datadir / 'collection'
+
+        def _rel(p, b):
+            return p.relative_to(b)
+
+        relative_dir = map(partial(_rel, b=dir), pathlib_walk(dir))
+        relative_expected = map(partial(_rel, b=expected),
+                                pathlib_walk(expected))
+        assert sorted(relative_dir) == sorted(relative_expected)
+
+    def test_three_part_bad_ver(self, datadir, tmpcwd, monkeypatch,
+                                requests_mock, invoker):
+        col_id = 'col11405'
+        col_version = '1.1'
+        trip_ver = '1.1.5'
+        col_uuid = 'b699648f-405b-429f-bf11-37bad4246e7c'
+        col_hash = '{}@{}'.format(col_uuid, '1.4')
+        base_url = 'https://archive.cnx.org'
+        metadata_url = '{}/content/{}/{}'.format(base_url, col_id, col_version)
+        extras_url = '{}/extras/{}'.format(base_url, col_hash)
+        extra_extras_url = '{}/extras/{}@{}'.format(base_url, col_uuid, '1.1')
+
+        # Register the data urls
+        for fname, url in (('contents_1.4.json', metadata_url),
+                           ('extras_1.4.json', extras_url),
+                           ('extras_old.json', extra_extras_url),
+                           ):
+            register_data_file(requests_mock, datadir, fname, url)
+
+        # Register bad version as 404
+        register_404(requests_mock,
+                     'https://archive.cnx.org/contents/'
+                     'b699648f-405b-429f-bf11-37bad4246e7c@1.5')
+
+        from nebu.cli.main import cli
+        args = ['get', 'test-env', col_id, trip_ver]
+        result = invoker(cli, args)
+
+        assert result.exit_code == 4
+
+        msg = "content unavailable for '{}/{}'".format(col_id, trip_ver)
+        assert msg in result.output
+
     def test_outside_cwd(self, datadir, tmpcwd, monkeypatch, requests_mock,
                          invoker):
         monkeypatch.chdir('/var')
@@ -413,4 +502,36 @@ class TestGetCmd:
         assert result.exit_code == 4
 
         msg = "content unavailable for '{}/{}'".format(col_id, col_ver)
+        assert msg in result.output
+
+    def test_failed_request_no_raw(self, datadir, requests_mock, invoker):
+        col_id = 'col11405'
+        col_version = 'latest'
+        col_uuid = 'b699648f-405b-429f-bf11-37bad4246e7c'
+        col_hash = '{}@{}'.format(col_uuid, '2.1')
+        base_url = 'https://archive.cnx.org'
+        metadata_url = '{}/content/{}/{}'.format(base_url, col_id, col_version)
+        extras_url = '{}/extras/{}'.format(base_url, col_hash)
+        contents_url = '{}/contents/{}'.format(base_url, col_hash)
+        raw_url = '{}/contents/{}?as_collated=False'.format(base_url, col_hash)
+
+        # Register the data urls
+        for fname, url in (('contents.json', contents_url),
+                           ('extras.json', extras_url),
+                           ):
+            register_data_file(requests_mock, datadir, fname, url)
+
+        requests_mock.get(metadata_url,
+                          status_code=301,
+                          headers={'location': contents_url})
+
+        register_404(requests_mock, raw_url)
+
+        from nebu.cli.main import cli
+        args = ['get', 'test-env', col_id, col_version]
+        result = invoker(cli, args)
+
+        assert result.exit_code == 4
+
+        msg = "content unavailable for '{}/{}'".format(col_id, col_version)
         assert msg in result.output