+++ nebuchadnezzar/docs/changes.rst
@@ -2,6 +2,8 @@
 Change Log
 ==========
 
+- Allow specifying minor version for `get`, with three-part version `1.X.Y`.
+
 8.0.5
 -----