+++ nebuchadnezzar/docs/changes.rst
@@ -2,6 +2,8 @@
 Change Log
 ==========
 
+9.0.0
+-----
 - Add switch to fetch all associated resources (images).
 - Allow specifying minor version for `get`, with three-part version `1.X.Y`.